Add new format FORMAT_EMPHASIS to handle <em> tags
