+++ aztec/src/main/kotlin/org/wordpress/aztec/spans/AztecStyleEmphasisSpan.kt
@@ -1,0 +1,18 @@
+package org.wordpress.aztec.spans
+
+import android.graphics.Typeface
+import org.wordpress.aztec.AztecAttributes
+
+class AztecStyleEmphasisSpan(attributes: AztecAttributes = AztecAttributes())
+    : AztecStyleSpan(Typeface.ITALIC, attributes) {
+
+    override val TAG by lazy {
+        when (style) {
+            Typeface.ITALIC -> {
+                return@lazy "em"
+            }
+        }
+        throw IllegalArgumentException()
+    }
+}
+
