+++ aztec/src/main/kotlin/org/wordpress/aztec/spans/AztecStyleItalicSpan.kt
@@ -4,15 +4,4 @@ import android.graphics.Typeface
 import org.wordpress.aztec.AztecAttributes
 
 class AztecStyleItalicSpan(attributes: AztecAttributes = AztecAttributes())
-    : AztecStyleSpan(Typeface.ITALIC, attributes) {
-
-    override val TAG by lazy {
-        when (style) {
-            Typeface.ITALIC -> {
-                return@lazy "em"
-            }
-        }
-        throw IllegalArgumentException()
-    }
-}
-
+    : AztecStyleSpan(Typeface.ITALIC, attributes)
